Two big features in this commit:
* Page editing permissions have been completely overhauled, a big win for large sites with many editors (#610)
* Slideshow slots now support displaying everything with certain categories and tags as an alternative to browsing for individual images (#628)

The New Page Editing Permissions
--------------------------------

We have a new approach to individual and group editing permissions for 1.5. The "editors and managers" distinction has been eliminated in favor of adding users and/or groups to a list of editors and checking an additional "add/delete pages" box for appropriate people (or groups).

Edit permissions are no longer inherited from ancestor pages. Instead, you can set them differently on every page. 

However, there are "apply to subpages" checkboxes to quickly apply any permission to all of the descendants of a page. Also, you can lock any page down so that only admins can edit it, and that setting is never clobbered by a cascade of permissions from a parent page.

New pages copy the permissions of the parent page.

The "edit admin lock" feature allows edit permissions to be overridden on a child page that is unusual, such as a page that displays faculty listings drawn from a database and should not be edited directly by department staff, even though they can edit the parent page.

apostrophe:migrate has been updated to add the necessary columns for the admin locks, and to run the new apostrophe:cascade-edit-permissions task once after adding them. This makes sure your existing 1.4 permissions on ancestor pages are copied to all child pages once so that your site initially behaves in the same way after the upgrade. It's possible to run that task more than once, but we recommend letting your permissions for subpages be independent and using the "apply to subpages" feature in the UI whenever desired.

In order to implement a single interface for adding editors to the site, the "candidate group" for the "manage" permission (aka "add/delete pages") has been merged with the candidate group for the edit permission. That is, if you are a candidate to edit you are also a candidate to manage (add/delete). It's still up to the admin to actually grant you that ability on a page, so this is just as secure as before.

A similar overhaul of view permissions is coming up next.

The new permissions editor needs styling (John is on it).

* * *

Fixed some conflicting relation names for aAccess and aGroupAccess.
aArray::listToHashById now works with arrays as well as objects (it fetches the 'id' element of arrays, ignoring the method parameter).




git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@2345 c89fab08-2f9b-42bf-8176-bd3ad6bee547

+++ lib/action/BaseaSlideshowSlotActions.class.php
@@ -11,16 +11,39 @@ class BaseaSlideshowSlotActions extends BaseaSlotActions
     }
     
     $this->editSetup();
-    $ids = preg_split('/,/', $request->getParameter('aMediaIds'));
-    $q = Doctrine::getTable('aMediaItem')->createQuery('m')->select('m.*')->whereIn('m.id', $ids)->andWhere('m.type = "image"');
-    // Let the query preserve order for us
-    $items = aDoctrine::orderByList($q, $ids)->execute();
-    $this->slot->unlink('MediaItems');
-    $links = aArray::getIds($items);
-    $this->slot->link('MediaItems', $links);
-    // Save just the order in the value field. Use a hash so we can add
-    // other metadata later
-    $this->slot->value = serialize(array('order' => $links));
-    $this->editSave();
+    
+    if ($request->hasParameter('aMediaIds'))
+    {
+      $ids = preg_split('/,/', $request->getParameter('aMediaIds'));
+      $q = Doctrine::getTable('aMediaItem')->createQuery('m')->select('m.*')->whereIn('m.id', $ids)->andWhere('m.type = "image"');
+      // Let the query preserve order for us
+      $items = aDoctrine::orderByList($q, $ids)->execute();
+      $this->slot->unlink('MediaItems');
+      $links = aArray::getIds($items);
+      $this->slot->link('MediaItems', $links);
+      // This isn't a normal form submission, but the act of selecting items for a
+      // slideshow implies we picked the 'selected' radio button, so just save 'form' as if
+      // that choice had been saved normally
+      $this->slot->value = serialize(array('form' => array('type' => 'selected'), 'order' => $links));
+      return $this->editSave();
+    }
+    else
+    {
+      // A normal form submission, used for the tags and categories option
+      $value = $this->getRequestParameter('slot-form-' . $this->id);
+      $this->form = new aSlideshowForm($this->id);
+      $this->form->bind($value);
+      if ($this->form->isValid())
+      {
+        $this->slot->setArrayValue(array('form' => $this->form->getValues()));
+        return $this->editSave();
+      }
+      else
+      {
+        // Makes $this->form available to the next iteration of the
+        // edit view so that validation errors can be seen, if any
+        return $this->editRetry();
+      }
+    }
   }
 }
